[Fix] GameResultList 리팩토링 및 피드백 반영 #357
