+++ components/game/big/GameResultBigScore.tsx
@@ -1,3 +1,4 @@
+import { getTimeAgo } from 'utils/handleTime';
 import styles from 'styles/game/GameResultItem.module.scss';
 
 interface GameResultBigScoreProps {
@@ -6,11 +7,6 @@ interface GameResultBigScoreProps {
   time: string;
   scoreTeam1?: number;
   scoreTeam2?: number;
-}
-
-interface ScoreViewProviderProps {
-  status: string;
-  time: string;
 }
 
 export default function GameResultBigScore({
@@ -22,7 +18,7 @@ export default function GameResultBigScore({
 }: GameResultBigScoreProps) {
   return (
     <div className={styles.bigScoreBoard}>
-      <ScoreViewProvider status={status} time={time} />
+      {makeScoreStatus(status, time)}
       <div className={styles.gameScore}>
         {mode === 'normal' ? 'VS' : `${scoreTeam1} : ${scoreTeam2}`}
       </div>
@@ -30,7 +26,7 @@ export default function GameResultBigScore({
   );
 }
 
-function ScoreViewProvider({ status, time }: ScoreViewProviderProps) {
+function makeScoreStatus(status: string, time: string) {
   switch (status) {
     case 'live':
       return <div className={styles.gameStatusLive}>Live</div>;
@@ -47,25 +43,4 @@ function ScoreViewProvider({ status, time }: ScoreViewProviderProps) {
     default:
       return null;
   }
-}
-
-function getElapsedTimeSeconds(gameTime: string) {
-  return (Number(new Date()) - Number(new Date(gameTime))) / 1000;
-}
-
-function getTimeAgo(gameTime: string) {
-  const elapsedTimeSeconds = getElapsedTimeSeconds(gameTime) - 60 * 10;
-  const timeUnits = [
-    { unit: '년', second: 60 * 60 * 24 * 365 },
-    { unit: '개월', second: 60 * 60 * 24 * 30 },
-    { unit: '일', second: 60 * 60 * 24 },
-    { unit: '시간', second: 60 * 60 },
-    { unit: '분', second: 60 },
-  ];
-
-  for (const timeUnit of timeUnits) {
-    const elapsedTime = Math.floor(elapsedTimeSeconds / timeUnit.second);
-    if (elapsedTime > 0) return `${elapsedTime}${timeUnit.unit} 전`;
-  }
-  return '방금 전';
 }
