+++ utils/handleTime.ts
@@ -42,3 +42,24 @@ export const isAfterMin = (gameTimeString: string, min: number) => {
   beforeMin.setMinutes(beforeMin.getMinutes() - min);
   return gameTime.getTime() <= beforeMin.getTime();
 };
+
+export const getElapsedTimeSeconds = (gameTime: string) => {
+  return (Number(new Date()) - Number(new Date(gameTime))) / 1000;
+};
+
+export const getTimeAgo = (gameTime: string) => {
+  const elapsedTimeSeconds = getElapsedTimeSeconds(gameTime) - 60 * 10;
+  const timeUnits = [
+    { unit: '년', second: 60 * 60 * 24 * 365 },
+    { unit: '개월', second: 60 * 60 * 24 * 30 },
+    { unit: '일', second: 60 * 60 * 24 },
+    { unit: '시간', second: 60 * 60 },
+    { unit: '분', second: 60 },
+  ];
+
+  for (const timeUnit of timeUnits) {
+    const elapsedTime = Math.floor(elapsedTimeSeconds / timeUnit.second);
+    if (elapsedTime > 0) return `${elapsedTime}${timeUnit.unit} 전`;
+  }
+  return '방금 전';
+};
